requirements.txt
aiofiles>=0.4.0
anyio>=1.2.1
async-generator>=1.10
attrs>=19.3.0
blinker>=1.4
blosc>=1.8.3; sys_platform != 'win32'
chardet>=3.0.4
Click>=7.0
Flask>=1.1.2
Flask-Cors>=3.0.8
gunicorn>=20.0.4
h11>=0.9.0
h2>=3.1.1
hpack>=3.0.0
httptools>=0.0.13
hyperframe>=5.2.0
idna>=2.8
itsdangerous>=1.1.0
Jinja2>=2.10.3
MarkupSafe>=1.1.1
msgpack==1.0.0
multidict>=4.7.2
numpy<=1.18.1
opencv-python>=4.1.2.30
outcome>=1.0.1
Pillow==6.0.0
priority>=1.3.0
python-dotenv>=0.10.3
Quart<=0.10.0
Quart-CORS<=0.2.0
requests>=2.22.0
requests-futures>=1.0.0
six>=1.13.0
sniffio>=1.1.0
sortedcontainers>=2.1.0
toml>=0.10.0
trio>=0.13.0
typing-extensions>=3.7.4.1
urllib3>=1.25.7
uvicorn<=0.11.5
uvloop>=0.14.0; sys_platform != 'win32'
websockets>=8.1
Werkzeug>=0.15.0
wsproto>=0.15.0
httpx==0.13.3
hypercorn>=0.5.4
grpcio==1.27.2
protobuf>=3.10.0
boto3>=1.9.66
pyyaml
fuzzywuzzy
GPUtil